divide connections into several files
